remote file download
In this example, the form takes a remote file URL as input, and when the "Download" button is clicked, the PHP code processes the input, validates the URL, downloads the file content using file_get_contents(), and then saves the content to a local file using file_put_contents().

Remember to create a folder named "downloads" in the same directory as the PHP file to store the downloaded files. Also, make sure that you understand the security implications of downloading files from remote sources and consider implementing additional security measures as needed.